rework on output

+++ sources/opc_func.cpp
@@ -108,7 +108,7 @@ uint16_t opc_update_uint16(string name, string str_type, uint16_t val)
   if (str_type == "f") {
     float fl = OPCs.getVar(n, fl);
     res = (uint16_t)(fl * 100);
-    OPCs.setVar(n, val * (float)0.01);
+    OPCs.setVar(n, (int16_t)val * (float)0.01);
   } else if (str_type == "i") {
     int16_t i16 = OPCs.getVar(n, i16);
     res = (uint16_t)i16;
